Paths in sketch tool
A literal SVG path can be specified as data for the sketch tool.  I could have done this using
xy data in an identical fashion to line graphs, but this doesn't help me get the freeform shapes I need to highlight areas on an image for hotspots.  I can get the svg directly out of illustrator, so I used that instead. Note that it doesn't get scaled, so is somewhat imperfect.