got jsdoc producing documentation for the js files that have jsdoc annotations.
That is all of the current bric files, although the comments can still use some cleaning up.

run "jsdoc -c jsdoc-conf.json ." from the js directory.
it will produce the documentation in an out subdirectory, the top of the documentation is index.html.

There are some issues w/ the current jsdoc in our files, due both to bugs in jsdoc and to my misunderstanding of how to specify some jsdoc attributes.

1) jsdoc is not correctly parsing doc comments immediately following another comment, to make it work I had to add a space between, ie this vim regex:
:%s'.\*//\*\*'*/ /**'

2) jsdoc is not handling the @constructor tag before the description, so move it after the description.

3) use the @link tag for url references